users/requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==3.14.0
django-dramatiq==0.11.2
django-redis==5.2.0
djangorestframework==3.14.0
dramatiq==1.14.2
Faker==18.6.2
gevent==22.10.2
greenlet==2.0.2
idna==3.4
pika==1.3.2
prometheus-client==0.17.0
psycopg==3.1.9
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
watchdog==3.0.0
watchdog-gevent==0.1.1
zope.event==4.6
zope.interface==6.0
